Changes for real hardware
I have gone through and added some things in order to make this function
on hardware. The code in apic.c to map a GSI to a given IRQ is
temporarily disabled. Some additional protections have been added to the
GDT to ensure that it actually managed allocate space for the container
and TSS. Lastly in smp.c, allocation of the high stack has been changed
to use the regular alloc instead of pmm_contig_alloc. This is because
contig_alloc shouldn't be relied on after allocations and frees have
already happened in a given freelist.