Remove SSE, fix up syscalls, utilize GSBase, pager
I have removed SSE, and separated out the PUSH/POP_ALL macros into a
separate context.asm file. I have fixed how syscalls work, now argument
registers (defined by the SYS-V ABI) are preserved until the handler
function is actually called. The page tables are still switched using a
global variable; however, it is more formally named and can be
standardized. The stack is now switched using a value stored in the
processor descriptor pointed to by KernelGSBase. The KernelGSBase has
been made use of on the BSP. For some reason it does not work on the
APs, but I'll sort that out later. The processor descriptor now also
stores a pl0_stack, which is the pointer to the start of the stack that
is to be switched to upon a syscall to privellege level 0. I am not
entirely sure why syscall doesn't use the TSS to change the stack, I
just know it doesn't. Lastly I have changed the clone function of the
pager a bit such that it will copy pages that are not mapped in the
original page tables, but it will not overwrite already existing
mappings in the copied to page tables.