Co-operate with changes in arch, working switcher
K/arch has been given superiority when it comes to defining the
standard, so removed all headers that define standards here and
integrated them in K/arch. I have also gotten switching for threads
working. I need to implement some locks such that there are no race
conditions when it comes to getting a thread from the process, as well
as going over all of this again. To do this: I have changed the GDT such
that it is now statically allocated into kernel memory, this is the case
for ISTs as well. The context registers have been updated to also push
CR3 to the stack, so some simplifications may be possible in
syscall.asm.