Scheduler hook, static stacks, syscalls
I have refined the scheduler hook in the timer interrupt more. This
should allow for the rotation of threads among the processors. By this I
mean if there are N processors and M threads where N>=M, then given a
suitable chunk of time to execute, the first N threads could be loaded,
then in the next slice of time the next N, and so on. A case does still
need to be handled: how is process->nextex reset? I have made the
syscall stacks also statically allocated in the kernel. Really
questioning whether this is the best course of action or if I should
sacrifice some simplicity in order to use dynamic memory. Lastly, I have
changed around how _syscall works. I think I can still improve it, but
it works fine for now.