Fix scheduler hook, co-operate with Karch
The scheduler hook has been fixed. The reason for the multiple tasks
executing at once was because if the hook found no next thread, it would
just skip the threading, and continue on with what it was already doing.
This is has now been resolved with Arc_ProcessorHold. Lastly, I just
made sure to implement the new functions that Karch defines.