Co-operate with arch
Now co-operates with arch with ProcessEntries, I have made some reforms to how get_thread works in the round robin scheduler, it will no longer return a NULL value as it now treats the first process as the hold process. I have made it so that syscall_4 now dequeues the current process, and that is working. Lastly, in order to not freeze up __asm__("sti"); has been moved to kernel.c from start.c (as to not freeze up due to how the new scheduler works)