SSE, syscall interface, memory troubles
I have re-introduced SSE, though it is still not used in the kernel. The syscall interface has been modified so that the typical SYS-V registers are used to pass the syscall's parameters. However, there is a major issue that is permiating throughout the kernel: accessing memory that is not yet mapped. For instance, to start a process, its stack needs to be set up with some initial values. This cannot be done as the memory that contains that stack is not yet mapped. I need to come up with some way to bridge the gap between userspace page tables and kernel space page tables such that the kernel is able to freely write to a process's page tables while the kernel remains isolated. I have attempted to do this making the pager's clone function directional and introducing some handling code on page fault to attempt to resolve this given an active process is running, but so far that has only yielded further errors.