Co-operate with arch, page table swapping
Changed start.c to co-operate with the introduction of the per process VMM. The syscall implementation no longer swaps the page tables. Instead, it relies on a page fault handler to map in more of the kernel to the current page tables. Of course, the permissions for the kernel pages are retained, but this should reduce paging overhead the longer a program runs as the parts of the kernel that the syscalls need will be already mapped in - requiring no switch to different paging tables.