smp: low memory allocations are too high
The allocations that are made in low memory are too high. Shrinking the
the limit does not fix this issue, so fixes need to be made to strictly
obey this limit in the PMM. To temporarily overcome this, the lowest
buddy structure has been moved to the front and the low memory
allocations made to be done in those sizes. Further, in smp.asm the
final call has been changed to a jmp to not take up stack space. A
triple fault is encountered after the statement to enable interrupts.
There needs to be a clean up of all of this code.