interrupt: swapgs, -early init, registry
The early init functions have been removed; these duties are now handled
by start.c. A macro has been defined to create naked function to call a
specified handler function to take care of an interrupt. This is the
function that should be registered not the handler function that is to
be called. Further, in the registry macro's assembly as well as in
idt.asm logic has been added such that the current code segment is
checked against the returning code segment. If the two values differ,
then a swapgs is called.