smp: current context's processor descriptor
A variable has been added, Arc_CurProcessorDescriptor, that uses the
internal architecture specific structure of a processor descriptor to
keep track of the current processor descriptor. This is done by giving
the processor the `__seg_gs` attribute, making accesses to the object
be done through an offset by the gs segment register.