context: changes to ordering, placement, init
Implement the init and uninit functions declared by K/arch. Changes have
been made as to the posistioning of CR0 and CR4 in the context. They
have now been moved into the frame portion of the context. This
simplifies swapping. RSP has also been removed from the GPR section of
the frame section. This is because RSP is already contained within frame.