apic.c: bad bad bad evil mistake
In my whimsy I threw out logic and wrote this if statement here. The
original idea was to keep the 8259 PIC enabled, and potentially
configure it, if no IO APICs existed on the system. This if statement
would, however, keep the 8259 PIC enabled even if there were IO APICs in
the system. I have removed commented it out as a temporary fix.