syscall/smp: fix stack underflow / mappings
The SMP implementation, upon registering a processor would fail to set
the syscall_stack value in the x64 processor descriptor to any value
causing errors upon invokation of a syscall. Further, the syscall
handler would naively set RSP to the syscall_stack value - or the base
of the stack. This would cause an underflow leading to further errors.
(The way I discovered this was that by some CHANCE syscall_stack was
being set to a value 256 bytes above the processor descriptor for
processor 0, for all other processors it would be 0, and that for some
reason after invoking two functions the last value of the x64 processor
descriptor, the one pointing to the abstract processor descriptor, would
corrupt and become a userspace stack pointer). A new function has been
added to the SMP that effectively maps in all of the processor specific
structures into a context that doesn't already have them.