syscall: userspace section, switching page tables
Since there are now two page tables, userspace-userspace and
userspace-kernelspace, for a given process; switching between them needs
to be implemented. In the syscalls this is acomplished by defining a
function to get the userspace-kernelspace page tables of the current
process. I have not implemented this very cleanly and would like to
revisit it. Further, the processor descriptor now also needs to keep
track of the current process and not just the current thread as the
current thread's descriptor cannot be cleanly mapped into the
userspace-userspace page tables.