smp/context: merge structs/implement functions
On the SMP side of things: the structures have been consolidated such
that the generic processor descriptor is now allocated inside of the
specific processor descriptor. The GDT, IDT, and TSS are now also
declared inside of the specific processor descriptor structure. For
context, I implemented the save and load functions, though not fully as
floating point state / other state is not saved nor restored with them.
Since the generic processor descriptor is now inside of the specific
one, I needed to rework the smp_get_processor descriptor function. I
unfortunately had to make it read the MSR value and cast that to the
processor descriptor instead of just using the variable for it. For this
I decided to add some additional functions. Though, I would really like
to cut down on the call depth for getting the generic processor descriptor.