Use new form of userspace delcaration
The macro for declaring something as part of userspace has changed to be
a little bit more complex. It asks for a specific subsection (i.e.
text). This is so that the compiler doesn't complain about variables and
functions being in the same section and enables me to, in the future,
separate them out if needed.