Move all CPUID checks to context
All the checks that use CPUID have been moved to context to be done in
the set_proc_features function. Further, the context saving and loading
functions have had the xsave/rstor and fxsave/rstor calls added to them.
I still need to determine how exactly to store all of the state as the
xsave space is variable in size barring the legacy fxsave region.
Further, the ARC_CHECK_FEATURE macro is now utilized.